Moved logic from ParentificationTransformer to NodeUtils.parentizeNode method
